+++ packages/react-native-eyecandy/src/components/Col/Col.tsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import { View, ViewProps, ViewStyle } from 'react-native';
+
+export interface ColProps extends ViewProps {
+  justifyContent?: ViewStyle['justifyContent'];
+  alignItems?: ViewStyle['alignItems'];
+}
+
+export function Col({ style, justifyContent, alignItems, ...props }: ColProps) {
+  return (
+    <View
+      style={[
+        {
+          alignItems,
+          justifyContent,
+        },
+        style,
+      ]}
+      {...props}
+    />
+  );
+}