+++ packages/react-native-eyecandy-icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nomada-sh/react-native-eyecandy-icons
 
+## 0.0.4
+
+### Patch Changes
+
+- Created Trash icon.
+
 ## 0.0.3
 
 ### Patch Changes