+++ packages/react-native-eyecandy/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @nomada-sh/react-native-eyecandy
 
+## 0.0.5
+
+### Patch Changes
+
+- Added `hidden` prop to Badge component.
+- Created Col component.
+- Created Row component.
+- Updated dependencies
+  - @nomada-sh/react-native-eyecandy-icons@0.0.4
+
 ## 0.0.4
 
 ### Patch Changes